[electron] Replace ffmpeg library
The ffmpeg library distributed alongside the Electron binaries contain
proprietary code for the h264 video codec. Fortunately enough, the
electron organization also releases a stripped ffmpeg library.

This commit adds a `postinstall` script to the `@theia/electron` package
to patch the pulled Electron binaries, and adds a test script playing a
video to check if the codecs are present or not.

The test script will exit with `code = 0` if the video cannot be played,
or with `code > 0` if anything else happens.

New scripts, to be ran using `npx <script>`:
- `electron-h264-test`
- `electron-replace-ffmpeg`

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>